# Improve com_content/models/articles.php to speed up big sites