@pierinz 2015-01-01
Finally, the answer.
Since we are using MySql on a test environment, after the first query the tables are in the mysql, ram and disk cache, and they will stay there for the whole test (as we recall the same data).
The patched and unpatched version yelds identical results when this data is cached. But the patched version use two queries (the unpatched has only 1) to perform the same operation **for each model call**, that's why it's slower (we can fix this).
The profiler shows different results because PHP loads lots of files to display the page, which may change the content of the disk cache.

On Linux there are some useful commands to erase ALL caches:
`sync`
`echo 3 > /proc/sys/vm/drop_caches`
`blockdev --flushbufs /dev/sda`
`hdparm -F /dev/sda`
then restart mysql service.

Results:
**DbForge Studio Profiler**
Unpatched: 2.702s
Patched: 0,088s + 2,219s

**Mysqlslap**
As we increase the iteration number, more data get cached and the time decreases.
Unpatched: mysqlslap --user=mysqlslap --password=mysqlslap --query=Desktop/unpatched.sql --delimiter=";" --concurrency 10 --iterations 10
Benchmark
        Average number of seconds to run all queries: 0.669 seconds
        Minimum number of seconds to run all queries: 0.214 seconds
        Maximum number of seconds to run all queries: 4.634 seconds
        Number of clients running queries: 10
        Average number of queries per client: 42

Patched: mysqlslap --user=mysqlslap --password=mysqlslap --query=Desktop/patched.sql --delimiter=";" --concurrency 10 --iterations 10
Benchmark
        Average number of seconds to run all queries: 0.628 seconds
        Minimum number of seconds to run all queries: 0.186 seconds
        Maximum number of seconds to run all queries: 4.502 seconds
        Number of clients running queries: 10
        Average number of queries per client: 48

**Conclusion**: on a shared hosting with many users we should get a huge improvement, with a dedicated server it should be almost the same. I should cache the result of the first query for all model calls in that request to decrease the queries.