@pierinz 2016-05-21
I can, but in this time I wondered if this is the right approach to solve the slowness.
Caching the badcategories and ignoring useless fields are good things, but the big deal is sorting the content table.

I thought we should have a flag in the module settings to enable "really big site mode".
In this mode the module should use an additional table "most_read_table" with "article id" and "hits" fields, which should contain only a small number of rows (the module's count parameter, for example).
When any "hits" field of the content table is updated, a mysql trigger (or some hook in the article) should check if the new value is higher than any of the "most_read_table" rows, then insert a new row with the "article id" and remove the row with the lowest hits.
I'm sure this will have a negative impact on the performances, but still should be faster than sorting the articles table.

Anyone have any thoughts on this?