@photodude 2016-05-21
@mbabker @wilsonge do either of you have thoughts on this?