@mbabker 2016-05-21
My honest opinion is that counter tables seem counter-intuitive to a system's design.

Also, database level triggers cannot be relied upon for a mass distributed application such as Joomla so any changes to implement this are going to have to be at the PHP layer, which is just going to add more processing time to article operations (especially page loads when hit counting is turned on as it's going to update both the content table and this "most_read_table").