@Bakual 2016-05-21
Imho, the solution to solve slow modules is to cache the module output itself. If you only have to fetch the module data and render the output once every few hours, then the whole thing speeds up a lot :smile: The good thing is, that feature is already there and can be turned on.
I certainly wouldn't add an own table for that module. That would be quite unique for a module.