@mbabker 2016-05-22
Cache is only a partial solution.  It helps speed up display of data, but for the request that builds the cache data it still needs to be optimized for the best performance we can give it.  If that means making it possible for the module to execute a more performant query with the articles model then it's a win for that module and anyone using the model to do lookups.

We threw a cache layer in front of the stats API because the base request for all data takes several seconds (it takes a lot of resources to process 375K rows of data) but once that cache is built for the next 15 minutes it's a pretty instantaneous return.  We also made performance tweaks like smaller queries (tradeoff being more had to be executed), shifting from returning `stdClass` objects to arrays (that did help a bit with the memory consumption, overall it allowed an additional iteration of our loop before hitting the server's memory limits), and ultimately using PHP 5.5's Generators (unfortunately that's not exactly an option here).  Same logic needs to apply here in as B/C a manner as practical.