@Bakual 2016-05-22
Yes sure, if the query can be optimised it's a no brainer to do it.