@PhilETaylor 2016-09-07
> Yes sure, if the query can be optimised it's a no brainer to do it.

We need to be removing the SECONDS from the publish_up/down times in sql queries... **been saying that for years**, because else the query changes every second and this is then not cached by mysql caching layers :( :(