@brianteeman 2016-09-07
>  *been saying that for years*,

So submit a pull request

On 7 September 2016 at 21:22, Phil Taylor <notifications@github.com> wrote:

> Yes sure, if the query can be optimised it's a no brainer to do it.
>
> We need to be removing the SECONDS from the publish_up/down times in sql
> queries... *been saying that for years*, because else the query changes
> every second and this is then not cached by mysql caching layers :( :(
>
> —
> You are receiving this because you commented.
> Reply to this email directly, view it on GitHub
> <https://github.com/joomla/joomla-cms/pull/5544#issuecomment-245404991>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/ABPH8dpwxWzqAU3qBNXALRjJIOXWm8Odks5qnx0hgaJpZM4DMsXk>
> .
>



-- 
Brian Teeman
Co-founder Joomla! and OpenSourceMatters Inc.
http://brian.teeman.net/