@csthomas 2016-09-07
What about remove badcats stuff at all and calculate path published state on category save event.

Take a look at #9160. 
It worked for 3.4 but it is uncompleted, idea should be good but requires a few changes.