@csthomas 2017-03-10
Please check competitive solution in PR #14439